// src/api/generated/models/CashierStats.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Статистика кассира за период
 */
export type CashierStats = {
  orders_completed: number;
  errors: number;
  rating: number;
  average_check: number;
  hours_processed: number;
  successful_orders_percent: number;
};